+++ src/app/api/admin/verify-otp/route.ts
@@ -1,40 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
-
-// Global OTP storage that persists across requests
-// In production, use Redis or database
-declare global {
-  var otpStorage:
-    | Map<
-        string,
-        {
-          code: string;
-          email: string;
-          createdAt: Date;
-          expiresAt: Date;
-          attempts: number;
-          isUsed: boolean;
-        }
-      >
-    | undefined;
-}
-
-const otpStorage =
-  globalThis.otpStorage ??
-  new Map<
-    string,
-    {
-      code: string;
-      email: string;
-      createdAt: Date;
-      expiresAt: Date;
-      attempts: number;
-      isUsed: boolean;
-    }
-  >();
-
-if (process.env.NODE_ENV !== "production") {
-  globalThis.otpStorage = otpStorage;
-}
+import {
+  getStoredOTP,
+  deleteStoredOTP,
+  updateStoredOTP,
+} from "@/lib/serverOtpStorage";
 
 // Get authorized emails from environment variable
 const getAuthorizedEmails = (): string[] => {
@@ -44,24 +13,6 @@ const getAuthorizedEmails = (): string[] => {
   }
   return emailsString.split(" ").filter((email) => email.trim().length > 0);
 };
-
-// Store OTP (called from send-otp route)
-export function storeOTP(email: string, otp: string): Date {
-  const now = new Date();
-  const expiresAt = new Date(now.getTime() + 5 * 60 * 1000); // 5 minutes
-
-  const key = email.toLowerCase().trim();
-  otpStorage.set(key, {
-    code: otp,
-    email,
-    createdAt: now,
-    expiresAt,
-    attempts: 0,
-    isUsed: false,
-  });
-
-  return expiresAt;
-}
 
 export async function POST(request: NextRequest) {
   try {
@@ -101,16 +52,13 @@ export async function POST(request: NextRequest) {
     }
 
     // Get stored OTP
-    const key = normalizedEmail;
-    const storedOTP = otpStorage.get(key);
+    const storedOTP = getStoredOTP(normalizedEmail);
 
     // Debug logging in development
     if (process.env.NODE_ENV === "development") {
       console.log("OTP Verification Debug:", {
         email: normalizedEmail,
         hasStoredOTP: !!storedOTP,
-        storageSize: otpStorage.size,
-        allKeys: Array.from(otpStorage.keys()),
       });
     }
 
@@ -132,7 +80,7 @@ export async function POST(request: NextRequest) {
     // Check if OTP is expired
     const now = new Date();
     if (now > storedOTP.expiresAt) {
-      otpStorage.delete(key);
+      deleteStoredOTP(normalizedEmail);
       return NextResponse.json(
         { error: "OTP has expired. Please request a new one." },
         { status: 400 }
@@ -145,7 +93,7 @@ export async function POST(request: NextRequest) {
     // Check if OTP matches
     if (storedOTP.code === otp.trim()) {
       // Mark as used and remove from storage
-      otpStorage.delete(key);
+      deleteStoredOTP(normalizedEmail);
 
       return NextResponse.json({
         success: true,
@@ -155,7 +103,7 @@ export async function POST(request: NextRequest) {
       const remainingAttempts = Math.max(0, 3 - storedOTP.attempts);
 
       if (remainingAttempts === 0) {
-        otpStorage.delete(key);
+        deleteStoredOTP(normalizedEmail);
         return NextResponse.json(
           {
             error: "Too many failed attempts. OTP has been invalidated.",
@@ -166,7 +114,7 @@ export async function POST(request: NextRequest) {
       }
 
       // Update the stored OTP with new attempt count
-      otpStorage.set(key, storedOTP);
+      updateStoredOTP(normalizedEmail, storedOTP);
 
       return NextResponse.json(
         {
